+++ src/main/java/com/syspa/login_service/security/SecurityConfig.java
@@ -76,7 +76,7 @@ public class SecurityConfig {
     .authorizeHttpRequests(
       registry -> {
         registry
-          .requestMatchers("/auth/*/signup", "/auth/*/login", "/auth/*/refresh")
+          .requestMatchers("/auth/*/signup", "/auth/*/login", "/auth/*/refresh", "/auth/*/logout")
           .permitAll();
         registry.requestMatchers("/actuator/health", "/actuator/metrics").permitAll();
         registry.requestMatchers("/.well-known/jwks.json", "/.well-known/openid-configuration").permitAll();